3. б)** Пересоздал библиотеку содержащую класс для работы с массивом. Продемонстрировать работу библиотеки. => Перенёс класс MyArray в библиотеку классов MyLibraryForArray и подключил эту библиотеку к текущему проекту. => Работа класса MyArray в текущем проекте демонстрирует работоспособность новой библиотеки.
