+++ C_Sharp_Level_1/C_Sharp_Level_1.sln
@@ -19,6 +19,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson4_PracticalPart", "Le
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson4_HomeWork", "Lesson4_HomeWork\Lesson4_HomeWork.csproj", "{E0DC4517-4BED-45F0-B98B-470F85F54ECA}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyLibraryForArray", "MyLibraryForArray\MyLibraryForArray.csproj", "{85D27AE4-99E2-4C51-82C8-B47D011A0640}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -57,6 +59,10 @@ Global
 		{E0DC4517-4BED-45F0-B98B-470F85F54ECA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E0DC4517-4BED-45F0-B98B-470F85F54ECA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E0DC4517-4BED-45F0-B98B-470F85F54ECA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{85D27AE4-99E2-4C51-82C8-B47D011A0640}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85D27AE4-99E2-4C51-82C8-B47D011A0640}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85D27AE4-99E2-4C51-82C8-B47D011A0640}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85D27AE4-99E2-4C51-82C8-B47D011A0640}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
