+++ C_Sharp_Level_1/Lesson4_PracticalPart/Program.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using static Lesson1_HomeWork.UsefulMethods;
+using MyLibraryForArray;
 
 namespace Lesson4_PracticalPart
 {
