+++ C_Sharp_Level_1/MyLibraryForArray/MyArray.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.IO;
 
-namespace Lesson4_PracticalPart
+namespace MyLibraryForArray
 {
     /// <summary>
     /// Библиотека для работы с массивом, включающим элементы типа int
@@ -118,7 +118,7 @@ namespace Lesson4_PracticalPart
                 {
                     sum += array[i];
                 }
-                return (double)sum / array.Length; 
+                return (double)sum / array.Length;
             }
         }
 
@@ -222,7 +222,7 @@ namespace Lesson4_PracticalPart
             string str = "";
             try
             {
-                
+
                 for (int i = 0; i < array.Length; i++)
                 {
                     str += $"{array[i]}, ";
@@ -246,7 +246,7 @@ namespace Lesson4_PracticalPart
             {
                 for (int j = 0; j < array.Length - 1; j++)
                 {
-                    if (array[j] > array[j + 1]) 
+                    if (array[j] > array[j + 1])
                     {
                         temp = array[j + 1];
                         array[j + 1] = array[j];
@@ -301,7 +301,7 @@ namespace Lesson4_PracticalPart
             int minOfArray = Min;
             foreach (var item in array)
             {
-                seriesN[item - minOfArray]++; 
+                seriesN[item - minOfArray]++;
             }
 
             int max = seriesN.Max;
@@ -313,7 +313,7 @@ namespace Lesson4_PracticalPart
 
             MyArray arrayOut = new MyArray(count);
             count = 0;
-            
+
             for (int i = 0; i < seriesN.Length; i++)
             {
                 if (seriesN[i] == max) { arrayOut[count] = minOfArray + i; count++; }
@@ -372,8 +372,85 @@ namespace Lesson4_PracticalPart
             Console.WriteLine("Элемент => Частота");
             foreach (var item in keys)
             {
-                Console.WriteLine($"{item, 7} => {frequencyItem[item]}");
+                Console.WriteLine($"{item,7} => {frequencyItem[item]}");
             }
         }
+
+        /// <summary>
+        /// Заполняет массив случайными числами в указанном диапазоне включительно
+        /// </summary>
+        /// <param name="array">Массив, который нужно заполнить случайными числами</param>
+        /// <param name="minValue">Минимальное значение для создания случайных чисел</param>
+        /// <param name="maxValue">Максимальное значение для создания случайных чисел</param>
+        public static void FillingArrayWithRandomNumbers(int[] array, int minValue, int maxValue)
+        {
+            Random rand = new Random(100);
+
+            for (int i = 0; i < array.Length; i++)
+            {
+                array[i] = rand.Next(minValue, maxValue + 1);
+            }
+        }
+
+        public static void OutputArray(int[] array)
+        {
+            Console.WriteLine("Выводим значения массива:");
+            for (int i = 0; i < array.Length; i++)
+            {
+                Console.Write($"{array[i]}, ");
+            }
+        }
+
+        /// <summary>
+        /// Возвращает количество пар элементов массива, в которых только одно число делится на указанный делитель.
+        /// Считаем, что ноль делится на любое число.
+        /// </summary>
+        /// <param name="array">Массив, пары элементов которых нужно проверить</param>
+        /// <param name="dividedBy">Делитель, на который нужно проверить делимость значений в парах элементов</param>
+        /// <returns>Возвращает количество пар элементов массива, в которых только одно число делится на указанный делитель.</returns>
+        public static int NumberOfPairsDividedBy(int[] array, int dividedBy)
+        {
+            //считаем, что ноль делится на любое число
+            int count = 0;
+            for (int i = 0; i < array.Length - 1; i++)
+            {
+                if (((array[i] % dividedBy == 0) && (array[i + 1] % dividedBy != 0))
+                    || ((array[i] % dividedBy != 0) && (array[i + 1] % dividedBy == 0))) count++;
+            }
+            return count;
+        }
+
+        /// <summary>
+        /// Заполняет значения элементов массива из файла (по одному числу с каждой следующей строки файла).
+        /// </summary>
+        /// <param name="array">Массив, который нужно заполнить из файла</param>
+        /// <param name="fileName">Путь к файлу, с указанием его имени</param>
+        public static void GetArrayFromFile(int[] array, string fileName)
+        {
+            if (File.Exists(fileName))
+            {
+                try
+                {
+                    StreamReader streamReader = new StreamReader(fileName);
+
+                    for (int i = 0; i < array.Length; i++)
+                    {
+                        if (streamReader.EndOfStream)
+                        {
+                            Console.WriteLine("В файле данные закончились раньше, чем элементы в массиве");
+                            break;
+                        }
+                        array[i] = int.Parse(streamReader.ReadLine());
+                    }
+                    streamReader.Close();
+                }
+                catch (FormatException ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
+            }
+            else Console.WriteLine($"В указанной директории файл не обнаружен: {fileName}");
+        }
+
     }
 }
